+++ api/writeAPIBlocking.go
@@ -7,7 +7,6 @@ package api
 import (
 	"context"
 	"strings"
-	"sync"
 
 	http2 "github.com/influxdata/influxdb-client-go/v2/api/http"
 	"github.com/influxdata/influxdb-client-go/v2/api/write"
@@ -61,7 +60,6 @@ type WriteAPIBlocking interface {
 type writeAPIBlocking struct {
 	service      *iwrite.Service
 	writeOptions *write.Options
-	lock         sync.Mutex
 }
 
 // NewWriteAPIBlocking creates new WriteAPIBlocking instance for org and bucket with underlying client
@@ -70,10 +68,11 @@ func NewWriteAPIBlocking(org string, bucket string, service http2.Service, write
 }
 
 func (w *writeAPIBlocking) write(ctx context.Context, line string) error {
-	w.lock.Lock()
-	defer w.lock.Unlock()
-	err := w.service.HandleWrite(ctx, iwrite.NewBatch(line, w.writeOptions.RetryInterval()))
-	return err
+	err := w.service.WriteBatch(ctx, iwrite.NewBatch(line, w.writeOptions.RetryInterval()))
+	if err != nil {
+		return err.Unwrap()
+	}
+	return nil
 }
 
 func (w *writeAPIBlocking) WriteRecord(ctx context.Context, line ...string) error {